pyproject.toml
[tool.poetry]
name = "scs-admin"
version = "0.1.0"
description = ""
authors = ["johnny.kim <bdkim855@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
fastapi = "^0.115.5"
uvicorn = {extras = ["standard"], version = "^0.32.0"}
python-dotenv = "^1.0.1"
python-multipart = "^0.0.17"
python-form = "^0.2.3"
fastapi-sqlalchemy = "^0.2.1"
jinja2 = "^3.1.4"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
pyjwt = "^2.10.0"
psycopg2-binary = "^2.9.10"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
